+++ include/bitcoin/bitcoin/message/get_address.hpp
@@ -43,9 +43,9 @@ public:
 
     void reset();
 
-    size_t satoshi_size() const;
+    uint64_t satoshi_size() const;
 
-    static size_t satoshi_fixed_size();
+    static uint64_t satoshi_fixed_size();
 };
 
 } // end message
